+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Fix control group attribute resolution so attributes of the same type (but associated with different pseudo states) won't be layered on top of each other.
 - Refactor the `StaticAttribute`/`ResponsiveAttribute`/`AnimatedAttribute` traits so value extraction can be customized.
 - Adjust field names of `AnimatedVals` and `AnimationSettings` to match the `Animated` instruction.
+- Add `UpdateId` system input type (like `In` or `Trigger`). This is used in `.update()` and `.update_on()` now instead of layered closures.
+- Add slider widget. Use the `Slider` and `SliderHandle` loadables to set up a slider. See the `slider` example.
 
 ## 0.5.0-rc.3